Essential routes
These are the routes that our app will be using mostly. As Typescript is a static typed language, so it is very essential to define the datatype of any parameter being used anywhere in the code. In Express.js there was no need to do that as Javascript is dynamically typed language. Therefor, we had to first import a definition of the typed, we were going to use here in the routes i.e. here used by the name of ctx and typename HttpContextContract. For any other type that is not defined in the language we must import that from @types/ so that we don't get any error. It is also very important as definition of parameters or types won't work otherwise.