Databases migrated to AWS
Instance on AWS RDS has been created and all these tables were migrated successfully to that instance. node ace migration plays a vital role in this process because that is responsible for that. Furthermore, it creates a tables for it's own use, where it keeps track of which of the files has been executed to create a table and which are left. SO, that when the next time if you run node ace migration:run command, it knows that which of the files are already migrated to the database. knex.js can be used to take help in creating schema for your database because that has all the necessary resources for almost all relational databases.