Model Relationships
First and foremost, it is very important to have a id in every table which is like a demand by Adonisjs or may be relational feature of database, but that is very essential otherwise there will be errors waiting. To have a relation you must create a column where you are going to relate your table. Only after that a relationship will be developed. For a 1-to-many or many-to-1 you have to create a hasMany in one table and belongsTo in other table. The belongsTo will contain the column name being pointed out by a Local Key so that the higher order table can identify where to push its data. And in higher order table, hasMany will have a foreign key be pointing out to the column of the dependent table so that it would know where to send its data. Record insertion can be achieved by the instance of the higher order table. Only that instance will be able to create the record in the right manner such that the relationship is maintained pointing to it.
Then there come the UUID as primary key and every table must have one. But assignment can only be done by a beforeCreate component. Or alternatively you can use integer being incremented own it's own but that is not an ideal scenario.